Return error stack or message.
Prior to this change, strinfigying an Error object just returns an empty object, so the error is not useful at all. This returns a more useful stack or message if the stack isn't available. We can potentially visually display stacks in the CLI and Inngest Cloud to make debugging issues easier for developers.